requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
# Add these when you implement the actual functionality:
# friendli
# pydub
# daft
# weaviate-client requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
weaviate-client==3.25.2
openai==1.3.5
pydantic==1.10.13
